+++ src/main/java/com/solvd/atm/persistence/impl/AccountRepositoryImpl.java
@@ -6,6 +6,8 @@ import com.solvd.atm.persistence.AccountRepository;
 import com.solvd.atm.persistence.MyBatisSessionHolder;
 import org.apache.ibatis.session.SqlSession;
 
+import java.math.BigDecimal;
+
 public class AccountRepositoryImpl implements AccountRepository {
 
     @Override
@@ -23,7 +25,7 @@ public class AccountRepositoryImpl implements AccountRepository {
     }
 
     @Override
-    public void changeAccountMoney(Account account, Integer money) {
+    public void changeAccountMoney(Account account, BigDecimal money) {
         try (SqlSession session = MyBatisSessionHolder.getSqlSessionFactory().openSession(true)) {
             session.getMapper(AccountRepository.class).changeAccountMoney(account, money);
         }
